+++ android/src/main/java/com/reactnativestripesdk/PaymentMethodMessagingElementConfig.kt
@@ -61,7 +61,7 @@ fun parseAppearance(
   font?.let { appearance.font(font) }
   val colors = PaymentMethodMessagingElement.Appearance.Colors()
   textColor?.let { colors.textColor(it) }
-  linkTextColor?.let { colors.infoIconColor(linkTextColor) }
+  linkTextColor?.let { colors.linkTextColor(linkTextColor) }
   appearance.colors(colors)
 
   return appearance